Finished Sign Up

+++ Back/src/main/java/com/example/spamclassifier/dto/MailDTO.java
@@ -1,12 +1,10 @@
 package com.example.spamclassifier.dto;
 
-import com.example.spamclassifier.model.User;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import javax.persistence.*;
 import java.time.LocalDateTime;
 
 @Getter
@@ -15,7 +13,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class MailDTO {
 
-    private Integer id;
+    private Long id;
     private String subject;
     private String message;
     private LocalDateTime createdAt;
